Update Entities and Gateway Events to V8
* Updates entities to V8
* Updates events to V8
* Make Snowflake a serializable type and introduce it to lower-level entities and core Data classes
* Introduce Optional delegates to ease porting REST requests and builders to V8
* Move the RestServiceTest to core and have it use core entities instead.

+++ core/src/test/kotlin/com/gitlab/kordlib/core/entity/GuildEmojiTest.kt
@@ -9,7 +9,7 @@ import mockKord
 internal class GuildEmojiTest : GuildEntityEqualityTest<GuildEmoji> by GuildEntityEqualityTest({ id, guildId ->
     val kord = mockKord()
     val data = mockk<EmojiData>()
-    every { data.id } returns id.longValue
-    every { data.guildId } returns guildId.longValue
+    every { data.id } returns id
+    every { data.guildId } returns guildId
     GuildEmoji(data, kord)
 })
